requirements.txt
cloudshell-cli~=6.0
cloudshell-l1-networking-core~=2.0
cloudshell-logging~=2.1
